requirements.txt
dagster == 1.3.7
dagit == 1.3.7
dagster-docker == 0.19.7
dagster-duckdb == 0.19.7
dagster-duckdb-pandas == 0.19.7
dagster-gcp == 0.19.7
openpyxl == 3.1.2
polars == 0.17.5
petl == 1.7.12
pyarrow == 11.0.0
xlsx2csv == 0.8.1
connectorx == 0.3.2a5
SQLAlchemy >= 1.0, < 2.0.0
requests == 2.31.0
google-cloud-bigquery == 3.11.0